bridge-utils: Fix compile against linux-3.8.x
Linux 3.8 has a header, include/uapi/linux/if_bridge.h that uses a
struct in6_addr but doesn't define it.  The trivial seeming fix of
including the header that does define it causes more problems.  The
problem was discussed on mailing lists in January 2013.  The final
suggestion I found was here:

      http://www.redhat.com/archives/libvir-list/2013-January/msg01253.html

This is intended to implement that suggestion.

Signed-off-by: Russell Senior <russell@personaltelco.net>